:recycle: Moved nearest marker getter to a new Map manager
